whisper/requirements.txt
fastapi
uvicorn
faster-whisper
requests
numpy
python-dotenv
